split tests to system dependent modules
